convert unstyled components to SFC
