+++ packages/vue/src/components/SignUpButton.vue
@@ -1,0 +1,50 @@
+<script setup lang="ts">
+import { useAttrs, useSlots } from 'vue';
+import type { SignUpProps } from '@clerk/types';
+import { useClerk } from '../composables/useClerk';
+import { assertSingleChild, normalizeWithDefaultValue } from '../utils';
+
+type SignUpButtonProps = {
+  unsafeMetadata?: SignUpUnsafeMetadata;
+} & Pick<
+  SignUpProps,
+  'fallbackRedirectUrl' | 'forceRedirectUrl' | 'signInForceRedirectUrl' | 'signInFallbackRedirectUrl' | 'initialValues'
+> & {
+    mode?: 'modal' | 'redirect';
+  };
+
+const props = defineProps<SignUpButtonProps>();
+
+const clerk = useClerk();
+const slots = useSlots();
+const attrs = useAttrs();
+
+function getChildComponent() {
+  const children = normalizeWithDefaultValue(slots.default?.({}), 'Sign up');
+  return assertSingleChild(children, 'SignUpButton');
+}
+
+function clickHandler() {
+  const { mode, ...opts } = props;
+
+  if (mode === 'modal') {
+    return clerk.value?.openSignUp(opts);
+  }
+
+  void clerk.value?.redirectToSignUp({
+    ...opts,
+    signUpFallbackRedirectUrl: props.fallbackRedirectUrl,
+    signUpForceRedirectUrl: props.forceRedirectUrl,
+  });
+}
+</script>
+
+<template>
+  <component
+    :is="getChildComponent"
+    v-bind="attrs"
+    @click="clickHandler"
+  >
+    <slot />
+  </component>
+</template>
